Fixes to attribute propagation issues for variables and types.
- A type declarations w/ attributes will propogate them to local
  and global variable declarations that use the type.  Addresses #248.

- A local or global variable declared with attributes will propogate
  them values that are later assigned to it.  Addresses #151, #67, #474.

Addresses #327 in that it should be a more consistent approach to
attribute propagation than before: the rules are that they can migrate
in one direction: type -> variable -> value.   It also clears up the
most common usage issue: the expectation that attributes in a local or
global variable declaration apply to the identifier in that declaration
such that later values assigned to it are subjected to those attributes.