Initial BinPAC++ interface for Bro.
It works!

    # ./configure --enable-debug --enable-hilti --with-hilti-config=/path/to/hilti-config

Bro looks in scripts/pac2 for *.pac2 files containing protocol grammars:

    # cat scripts/pac2/ssh.pac2

    module SSH;

    export type Banner = unit {
        magic   : /SSH-/;
        version : /[^-]*/;
        dash    : /-/;
        software: /[^\r\n]*/;
    };

Bro also looks in scripts/pac2 for *.evt files that describe
(1) analyzers that use the *.pac2 grammars, and (2) events that the
parsing of those grammars are to trigger:

    # cat scripts/pac2/ssh.evt

    analyzer SSH over TCP:
        parse with SSH::Banner,
        port 22/tcp;

    on SSH::Banner -> event ssh::banner($conn, $is_orig, version, software);

That's actually all that's needed for defining a new analyzer with
corresponding events. Bro then compiles these analyzers at startup and
uses them right away for parsing once a connection on the specified
port comes along:

    # cat ./ssh-banner.bro

    event ssh::banner(c: connection, is_orig: bool, version: string, software: string)
        {
        print "SSH banner", c$id, is_orig, version, software;
        }

    # bro -r ssh-single-conn.trace ./ssh-banner.bro
    SSH banner, [orig_h=192.150.186.169, orig_p=49244/tcp, resp_h=131.159.14.23, resp_p=22/tcp], F, 1.99, OpenSSH_3.9p1
    SSH banner, [orig_h=192.150.186.169, orig_p=49244/tcp, resp_h=131.159.14.23, resp_p=22/tcp], T, 2.0, OpenSSH_3.8.1p1

Note that this is all barely more than a proof-of-concept right now; I
actually haven't tried it beyond this toy example yet. Only string
parameters are supported for now, there's almost certainly memory
trouble with traces of any significant size, and performance is
probably pretty bad currently as well.