Adding gzip/deflate support to HTTP.
The generated HTTP log is now quite similar to what a stock Bro
generates, including with full hashing turned on. There are still a
few differences in the size and hash fields that I haven't tracked
down yet, but generally it seems to work well. It's quite a bit slower
though, will do some profiling next.