Removing Broxygen's dpd_config magic.
The table doesn't exist anymore. The functionality has been replaced
with a function call but I can't really see how to get it out of
there. We could get it from the analyzer manager, however then we
can't tie it back to a script anymore. What we could do eventually is
add this to the information about the analyzer plugin.