Extending the plugin API to allow specifying extensions of input files that the plugin handles.
A plugin can now specify a set of file extensions that it can handle.
If Bro attempts to load such as file (e.g., from the command line),
instead of feeding it to the script interpreter as it would normally
do for a Bro script, it will pass it off to the plugin to process and
otherwise ignore.

Also renaming BROPLUGINS to BRO_PLUGINS.