temporary osx mavericks libc++ issue workaround.
When using clang3.3 on mavericks with libc++, after the
first std::getline-call encountered eof on an ifstream,
all following std::getline calls on the same ifstream will
also get eof, even if ifstream::clear has been called and
the file has been appended in the meantime.

Seeking to the current position after a clear before trying
to read the line fixes this behavior.

This is just a temporary workaround...