Fix the dir module.
Internally, Dir kept track of the files in directory
by storing all inode numbers in a set. However, when a file
is deleted and a new file is created in a directory, the old
file may get the same inode number as the old one. In this
case, bro did not notice the new file.

The patch simply changes the indexing of files - now files
are indexed by inode and creation time.

This should fix the scripts.base.utils.dir test failures.