Change connection role flipping logic.
Roles are now flipped immediately for connections that start with a
syn-ack packet instead of also waiting for a syn.  This means roles are
flipped at an earlier point in time and so there's less potential for
internal state to become invalid (e.g. with the old method, connection
history strings would still show originator activity after the role
flip).

Also noticed that the packet-wise child analyzers of the TCP analyzer
did not get notified via FlipRoles.